Use a reflex to fetch products instead of using controller
We can then load products on page load, from client side. This allows us to manage loading state (it could be relatively long to load products)

Basically, this commit:
 - Moves the table into its own partial
 - Moves methods to fetch products into a reflex, and delete them from controller
 - Creates a stimulusjs controller for productsV3 page. This will help us to manage state.
